requirements.txt
cmake
build
setuptools
wheel
numpy
dlib==19.24.6 --prefer-binary
streamlit
pillow
requests
replicate
python-dotenv
openai
matplotlib
opencv-python-headless
reportlab
numpy
dlib
fpdf2
fonttools
plotly
kaleido